packages/vtuber-behavior-engine/pyproject.toml
[project]
name = "vtuber_behavior_engine"
version = "0.1.0"
description = "AI core for the VTuber system, handling dialogue and behavior generation using ADK."
authors = [
    { name = "LongbowXXX" },
]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "python-dotenv>=1.2.1",
    "google-adk>=1.21.0",
    "google-genai>=1.56.0",
    "google-cloud-speech>=2.35.0",
    "pyaudio>=0.2.14",
    "mcp[cli]>=1.25.0",
    "wheel>=0.45.1",
    "setuptools>=80.9.0",
    "litellm>=1.80.11",
    "google-api-python-client>=2.187.0",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.3.0",
    "chromadb>=1.4.0",
    "google-generativeai>=0.8.6",
    "pyyaml>=6.0.2",
]


[project.optional-dependencies]
dev = [
    "types-requests>=2.32.4",
    "black>=25.12.0",
    "flake8>=7.3.0",
    "flake8-copyright>=0.2.4",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-html>=4.1.1",
    "respx>=0.21.1",
    "sphinx>=8.1.3,<9.0.0",
    "sphinx-rtd-theme==3.0.2",
    "myst-parser==4.0.1",
    "sphinxcontrib-mermaid==1.2.3",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
pythonpath = ["."]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] [%(thread)d] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
asyncio_default_fixture_loop_scope = "function"
generate_report_on_test = true
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.coverage.run]
branch = true
omit = ["tests/*"]

[tool.mypy]
plugins = []
ignore_missing_imports = true
strict = true
follow_imports = "normal"
strict_optional = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
show_error_codes = true
warn_unused_ignores = true
show_error_context = true
show_column_numbers = true

[[tool.mypy.overrides]]
module = "google.*" # Ignore missing stubs for Google libraries if needed
ignore_missing_imports = true
disallow_any_unimported = false

[tool.black]
target-version = ["py310"]
line-length = 120

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
